+++ dishrec_esp32/src/iXML.h
@@ -462,11 +462,11 @@ public:
     virtual void clear_umid();
 
     // Loudness
-    virtual void set_loudness_value(uint16_t value);
-    virtual void set_loudness_range(uint16_t range);
-    virtual void set_loudness_max_true_peak(uint16_t level);
-    virtual void set_loudness_max_momentary(uint16_t level);
-    virtual void set_loudness_max_short_term(uint16_t value);
+    virtual void set_loudness_value(double value);
+    virtual void set_loudness_range(double range);
+    virtual void set_loudness_max_true_peak(double level);
+    virtual void set_loudness_max_momentary(double level);
+    virtual void set_loudness_max_short_term(double value);
     virtual void clear_loudness();
 
     // Reserved
